+++ HISTORY.rst
@@ -13,6 +13,7 @@ Unreleased
 * ALCN-967: add payment mode "Domiciliation" and do the mapping
 * ALCN-973: add mapping for packaging quantities (box, pallet, shrink wrapping)
 * ALCN-989: automatize hot DB2 data import for INT and PROD
+* ALCN-981 - Set fridge route on product stocked in Q sector
 
 **Features and Improvements**
 * ALCN-946: Add tags on itinerary to easily target customers